Reviewed, added and rejigged code_documentation content (#35)
* Update the coding docs

This more or less finishes the first pass of the rewrite of the `code
documentation` section. I'm yet to read it and actually think of what
else needs to be covered. No external (to me) comments have been
provided on this commit and branch as of yet.

* Second pass, still no comments

This commit has a second pass over the docs, tweaking and adapting
sections as well as making more use of the `{note}` directive to format
notes.

* Add phils comments on code commenting

* Take onboard all comments from @philip-lee-ons

All the comments are now in place. Happy for @iangrimstead to review
now.

* Added notes regarding out-of-date comments and long functions

- long funcs can be split into a "root" func calling helpers; this expresses control flow and replaces comments with function names
- notes on code is the point of truth vs comments
...feels like a set of notes on structure wouldn't go amiss, but thats system design really...

* Update code_documentation.md

* Fix spelling

This fixes spelling mistakes I could find in the Code Documentation
section in preparation for the PR to be merged to the main Best Practice
Team repo

Co-authored-by: IanGrimstead <38883454+IanGrimstead@users.noreply.github.com>