Rework Core programming (#40)
* Core programming (#38)

* Update the coding docs

This more or less finishes the first pass of the rewrite of the `code
documentation` section. I'm yet to read it and actually think of what
else needs to be covered. No external (to me) comments have been
provided on this commit and branch as of yet.

* Second pass, still no comments

This commit has a second pass over the docs, tweaking and adapting
sections as well as making more use of the `{note}` directive to format
notes.

* Add phils comments on code commenting

* Take onboard all comments from @philip-lee-ons

All the comments are now in place. Happy for @iangrimstead to review
now.

* Added notes regarding out-of-date comments and long functions

- long funcs can be split into a "root" func calling helpers; this expresses control flow and replaces comments with function names
- notes on code is the point of truth vs comments
...feels like a set of notes on structure wouldn't go amiss, but thats system design really...

* Update code_documentation.md

* Add Functions section

Adds the functions section on what nice functions are to the code
modularisation chapter. Need @iangrimstead to look at and write the
class equivalent of this.

* Add a few bits on notebooks

* Add more context around modularised analysis

This commit fills out the section on modularised analysis and the
notebook sections. I think these are more or less ready for review to be
fair.

* Add more info on Modules

This commit adds the first draft of the modules section

* Tweak some formatting

* Spelling mistake

* Tweaks to structure before addressing clean code

Went through again, rejigged the structure a little bit and added extra
content and spelling/grammar/language fixes.

* Add tons of work on code style and naming

Currently stopped after the write up of idiomatic code.

* Fix a code example - small tweak

* Fix markdown italics

* Spelling

* Tweak chapter on automatically styling your code

This tweaks the chapter on linting and formatting. Adding a bit more
context and covering off pre-commit hooks

* Fix spelling

This fixes spelling mistakes I could find in the Code Documentation
section in preparation for the PR to be merged to the main Best Practice
Team repo

* Start addressing the KISS and SOLID sections

This so far is just a write up of the Simplicity point.

* Tweak Simplicity section

Just a few wording and narrative tweaks.

* Finish repetition section

This section contains mostly comments on repetition with code examples
of what not to do. I've review this and tweaked the Python side of the
code examples as well as added a bit more content.

* First class section drafted

* Few words on class naming

* Reviewed up to "Scripts"

* Added a few more bits, extra typos fixed too

* First draft

* Spelling and grammar

Also some tweaks to the formatting.

Add more changes to the SOLID bit

This basically removes the explicit references to the SOLID section and
it adds it all into the Software concepts for analysts portion.

* Remove most bolding

As requested by @alexander-newton I've removed bolding for readability purposes.

* Addressing more comments from @alexander-newton

Co-authored-by: IanGrimstead <38883454+IanGrimstead@users.noreply.github.com>
Co-authored-by: Ian Grimstead <Ian.Grimstead@ons.gov.uk>

* Separate learning resource and rearrange toc

* Add remaining resources to learning file

* Add pre-requisites

* Working through functions and classes

* Working through classes and clean code

* Finish working through core

Co-authored-by: Art <15175738+finnkauski@users.noreply.github.com>
Co-authored-by: IanGrimstead <38883454+IanGrimstead@users.noreply.github.com>
Co-authored-by: Ian Grimstead <Ian.Grimstead@ons.gov.uk>