Typos and style guide/accessibility checks
Didn't think the epigraph formatting was very accessible with the scroll bar and the aqua book  link written out in full so changed to Quote markup. 
Think the last section needs a subheading - I added in "Case Studies" - but think the structure could be improved a bit - I got a bit lost when you walked about continuous integration - couldn't tell if this was something PHE had done or something you were mentioning in general? 
Maybe more of a bullet point list of what happened in the case studies? 
Finally there seemed to be a link missing for "continuous integration" so I added one in.