typos and accessibility
Don't use italics, if you want to draw attention the odd word here and there - use bold as it is more accessible. 
I don't think that first image needs alt text - you can just mark it as decorative - I have changed the alt text to alt text="" which should do this - but I am not 100% sure how this works with markdown (personally I don't think you need that image at all -  I am not sure it is adding much and I did think it looked like the top had been mistakenly cropped off)
I am not sure about the formatting for the "Key strategies" section. I am not sure if this is something you use throughout the book, but if it isn't used that often it kind of disrupts the flow and the consistency? I can't see anything wrong with it from an accessibility point of view though - although you can never be sure - which is why I generally like to keep things simple! : )