typos and style guide
Unsure about the cartoon - you will have to check you can use it from a legal point of view - also there are accessibility issues when it comes to images of text. 
Also - the images (?) of code use italics - I guess this replicates what the code looks like in real life so the italics are OK in this instance. Are these sections images of text though? I don't know how we are going to get round that. Images of text fail accessibility regs. Unless they are supplied in an alternative format perhaps. 
Later on, in the separate responsibilities section there is another image of text, which I think is really a table - again, strictly speaking this fails regs, could it be presented as a table? Is it definitely needed? The alt text doesn't tell you the text in the image.