accessibility, typos and style guide edits
Just realised - a lot of text is coloured - have you check the contrast between the colour of the text and the background e.g. .R and .py are in that pink colour. Not sure about contrast with the white. The WebAIM contrast checker is a good one: https://webaim.org/resources/contrastchecker/

Also be good to double check your hyperlink text colour - as it is not automatically underlined it comes under stricter conditions about the contrast it has to have with the surrounding text. The WebAIM hyperlink text checker can tell you more about this: https://webaim.org/resources/linkcontrastchecker/

I would mark the DAG diagram as decorative or the text the screenreader is reading out is a bit repetitive. Not sure how to do this on github, normally you just replace the alt text with ""