typos and style guide
I would keep quotes to one style and then key points (like the one about comments or the "walks like a duck" one in core programming) in another style. To be consistent. 

Also I would check that all your hyperlinks have unique descriptive text - a screen reader user will often as a screen reader to read out all the links on a page as a way of "scanning" the content - think whether your link text would make sense to a person doing this. I think some of them need a bit of editing.